converted util functions to typescript and relocated the file
