+++ taxonomy-editor-frontend/src/pages/allentries/index.jsx
@@ -29,7 +29,7 @@ import DialogTitle from "@mui/material/DialogTitle";
 import Select from "@mui/material/Select";
 import ISO6391 from "iso-639-1";
 import { createBaseURL } from "../editentry/createURL";
-import { toTitleCase } from "../../components/interConvertNames";
+import { toTitleCase } from "../../utils";
 import { greyHexCode } from "../../constants";
 
 const Entry = ({ setDisplayedPages }) => {
